Move tile squareSearch algorithm into its own header
